Minor script addition to simplify automation
Hi Robert, this is a tiny change to build_kernel.sh to simplify it's integration into a larger meta-build system. All it does is to write the kernel version and local build tag (if that is what you call boneXX etc) to a kernel_version file so the appropriate wire-ups can be made further up the food chain. BR.